+++ internal/transport/http/DTO/clawMachine_DTO.go
@@ -20,11 +20,11 @@ type CreateClawItemsRequest struct {
 
 // CreateClawItemRequest represents the HTTP request for creating a single claw item
 type CreateClawItemRequest struct {
-	Name            string `json:"name" binding:"required"`
-	Rarity          string `json:"rarity" binding:"required"`
-	SpawnPercentage int64  `json:"spawnPercentage" binding:"required"`
-	CatchPercentage int64  `json:"catchPercentage" binding:"required"`
-	MaxItemSpawned  int64  `json:"maxItemSpawned" binding:"required"`
+	Name            string  `json:"name" binding:"required"`
+	Rarity          string  `json:"rarity" binding:"required"`
+	SpawnPercentage float64 `json:"spawnPercentage" binding:"required"`
+	CatchPercentage float64 `json:"catchPercentage" binding:"required"`
+	MaxItemSpawned  int64   `json:"maxItemSpawned" binding:"required"`
 }
 
 type CreateClawPlayerRequest struct {
